Move handling of missing file to normal code path
As stated in the comments of the original code, handling the missing
file using an exception is not appropriate, because it is a normal
situation (e.g. when starting the program for the very first time).
Therefore, the logic should be handled outside the exception code path
to aid readability.