Add checkstyle to gradle
While checkstyle is not covered as a learning objective in level 2, its
absence makes it harder for maintainers of the repository to ensure that
the coding standards are followed, as some violations (such as trailing
whitespaces and redundant imports) can be difficult to spot unless one
pays a lot of attention to them.

checkstyle.xml is taken from level 4.