Generate gradle wrapper
The gradle wrapper had to be generated as Travis was not able to execute
the checkstyle tasks without facing the following error:

    Execution failed for task ':checkstyleMain'.
        > java.lang.ClassNotFoundException:
        > com.puppycrawl.tools.checkstyle.CheckStyleTask

There were some breaking changes made to checkstyle that causes
incompatibility with older versions of Gradle. (more info here:
https://github.com/checkstyle/checkstyle/issues/2107)

At the time of writing, Travis uses Gradle 2.2.1.

In order to resolve this issue (and to ensure a standard gradle
version used by everyone), the wrapper is generated.