Generate gradle wrapper
The gradle wrapper have to be generated as Travis is not able to
execute the checkstyle tasks without facing the following error:

    Execution failed for task ':checkstyleMain'.
        > java.lang.ClassNotFoundException:
        > com.puppycrawl.tools.checkstyle.CheckStyleTask

There were some breaking changes made to checkstyle that causes
incompatibility with older versions of Gradle, such as Gradle 2.2.1,
which is used by Travis. (more info about the breaking changes here:
https://github.com/checkstyle/checkstyle/issues/2107)

In order to resolve this issue (and to ensure a standard gradle
version used by everyone), generate the wrapper.