AddressBook: remove unused addTag()
AddressBook#addTag() adds a Tag to the master list of Tags in
AddressBook.

This method is not used anywhere, and can be classified as unused code.
There are no commands that support the addition of Tags only, and it is
also not part of any learning objectives. Furthermore,
AddressBook#syncTagsWithMasterList() already handles the addition of
Tags from a Person gracefully without using exceptions. It should be
removed as unused code contributes to maintenance and comprehension
overheads.

Let's remove AddressBook#addTag().