AddressBook: remove unused addTag()
AddressBook#addTag() adds a Tag to the master list of Tags in
AddressBook. However this method is not used anywhere as there are no
commands that support the addition of Tags only; is not part of any
learning objectives; and AddressBook#syncTagsWithMasterList() already
handles the addition of Tags from a Person gracefully without using
exceptions.

Let's remove AddressBook#addTag() as unused code contributes to
maintenance and comprehension overheads.