AddressBook: remove unused removeTag()
AddressBook#removeTag() removes a Tag from the master list of Tags and
from any Person in AddressBook.

This method is not used anywhere, and can be classified as unused code.
There are no commands that support the removal of Tags only, and it is
also not part of any learning objectives. It should be removed as unused
code contributes to maintenance and comprehension overheads.

Let's remove AddressBook#removeTag().