AddressBook: remove unused removeTag()
AddressBook#removeTag() removes a Tag from the master list of Tags and
from any Person in AddressBook. However this method is not used anywhere
as there are no commands that support the removal of Tags only, and is
not part of any learning objectives.

Let's remove AddressBook#removeTag() as unused code contributes to
maintenance and comprehension overheads.