AddressBook: remove unused containsTag()
AddressBook#containsTag() checks if a given Tag is in the master list of
Tags. However this method is not used anywhere as there are no commands
that support it. It is also not part of any learning objectives. While
it is used in two tests for AddressBook#addPerson(), it can be replaced
with the existing AddressBookTest#isTagObjectInAddressBookList().

Let's remove AddressBook#containsTag() as unused code contributes to
maintenance and comprehension overheads.