AddressBook: remove unused containsTag()
AddressBook#containsTag() checks if a given Tag is in the master list of
Tags. However this method is not used anywhere as there are no commands
that support it, and is not part of any learning objectives. While it is
used in two tests for AddressBook#addPerson(),
AddressBook#containsTag(Tag):boolean is used infrequently enough to
justify superseding it with AddressBook#getAllTags():UniqueTagList and
UniqueTagList#contains(Tag).

Let's remove AddressBook#containsTag() as unused code contributes to
maintenance and comprehension overheads.