AddressBook: remove unused containsTag()
AddressBook#containsTag() checks if a given Tag is in the master list of
Tags.

This method is not used anywhere, and can be classified as unused code.
There are no commands that support it, and it is also not part of any
learning objectives. While it is used in two tests for
AddressBook#addPerson(), it should be replaced with the existing
AddressBookTest#isTagObjectInAddressBookList() to improve code reuse.
AddressBook#containsTag() should be removed as unused code contributes
to maintenance and comprehension overheads.

Let's remove AddressBook#containsTag().