UniqueTagList: remove unused add()
UniqueTagList#add() adds a Tag to the list of unique Tags, and throws an
exception if it is a duplicate of an existing Tag in the list. With the
removal of AddressBook#addTag(), it is now unused code.

Let's remove UniqueTagList#add() as unused code contributes to
maintenance and comprehension overheads.