UniqueTagList: remove unused remove()
UniqueTagList#remove() removes a Tag from the list of unique Tags, and
throws an exception if it is not found in the list.

With the removal of AddressBook#removeTag(), this method is not used
anywhere, and can be classified as unused code. It should be removed as
unused code contributes to maintenance and comprehension overheads.

Let's remove UniqueTagList#remove().