runtest.sh: fix bug in deleting actual.txt from previous run
There are three bugs in deleting actual.txt from previous run.

First, we should delete actual.txt when it exists, not it doesn't
exist. Thus we should not add "!" in the test command.

Second, actual.txt is a file and its existence should be tested
using "-e" not "-d".

Third, its directory should be "." not "..".

Let's fix these bugs.