DeveloperGuide: Improve project setup instructions
If the project is not set up properly, StorageFileTest will fail because
it cannot find the test data file that it uses.

The instructions in DeveloperGuide are not clear enough, as the
resolution of this problem is only in the Troubleshooting section, and
is not mentioned anywhere else.

Let's make the DeveloperGuide instructions clearer by addressing this
issue directly in the setup section.