Setup asciidoc-to-html build system
Our markdown documentation is published to GitHub pages via its
automatic markdown files publishing support[1]. However, this mechanism
does not support asciidoc files.

We are in the process of converting all of our documentation to
asciidoc, so we'll need to setup our own GitHub pages build and
deployment system.

Let's reuse the asciidoc-to-html build system in addressbook-level4[2]
by copying the relevant code/files.

[1] https://github.com/blog/2289-publishing-with-github-pages-now-as-easy-as-1-2-3
[2] https://github.com/se-edu/addressbook-level4