Upgrade project to run on Java 9
The project currently runs on Java 8.

Since AddressBookL4 is already running on Java 9, this project should
be upgraded to run on Java 9 as well for consistency.

Let's upgrade the project to run on Java 9 and update the
documentation accordingly.

Travis builds the project using JDK 8. Since we are now running on
Java 9, let's update it to build the project using JDK 9.

Java 9 introduces a module system, causing internal packages such as
javax.* and com.sun.* packages to be inaccessible[1]. As our project
uses these packages, it is no longer compilable. To resolve this,
let's:

- Update the documentation to mention how to import the
JAXB module, and adding of the '--add-module java.xml.bind' VM flag,
in order to build and run the project in Intellij or in Command Line.
- Update runtests.sh and runtests.bat to set the
'--add-module java.xml.bind' VM flag for compiling and running of the
test.
- Update build.gradle to import these modules, which is used in Travis.

[1]: http://openjdk.java.net/jeps/260