AdaptedAddressBook: remove master tag list
AdaptedAddressBook keeps a list of tags, which is used for
storing the master tag list in the AddressBook object.

Since AddressBook no longer has a master tag list, it is unnecessary
to store the master tag list in AdaptedAddressBook.

Let's remove the master tag list in AdaptedAddressBook.

Similarly, it is now unnecessary for AddressBook to have a
method for retrieving the master tag list.

Let's remove AddressBook#getTagList().