docs: update component on UniqueTagList
AddressBook previously had a UniqueTagList linked to it as the
master tag list, which was reflected in the architecture diagrams
in the DeveloperGuide and LearningOutcomes.

Since AddressBook does not use a master tag list anymore, and
UniqueTagList has been removed, these architecture diagrams are no
longer consistent with the program.

Let's update these architecture diagrams in the DeveloperGuide
and LearningOutcomes.