Storage: add AddressBookEncoder and AddressBookDecoder classes
Storage uses the JAXB module for storing and retrieving of the
address book data in a file.

However, we are trying to ensure Java 9 compatibility in #381,
but the JAXB module is deprecated from Java 9 onwards.

As use of deprecated modules is considered bad practice, we should
replace our storage needs with another library before moving on
to ensure Java 9 compatiblity.

As suggested here[1], let's reuse the storage code from
addressbook-level1.

To do this, let's create new classes, AddressBookEncoder and
AddressBookDecoder, which contains the ported storage code from
addressbook-level1, and update StorageFile to use them.

[1]: https://github.com/se-edu/addressbook-level2/pull/382#issuecomment-409441789