DeveloperGuide: update project to use Java 9
The project currently runs on Java 8.

Since AddressBookL4 is already running on Java 9, this project should
be upgraded to run on Java 9 as well for consistency.

Since the project has been adjusted to be Java 9 compatible in #388,
let's update the DeveloperGuide to mention using Java 9 or later.

Travis currently builds the project using JDK 8. Since we are now
running on Java 9, let's update it to build the project using JDK 9
as well.

UtilsTest uses the Integer(int) constructor for its equality test,
which is deprecated from Java 9[1]. Since use of deprecated APIs is
considered bad practice, let's replace this with Integer.valueOf(int).

[1] https://docs.oracle.com/javase/9/docs/api/java/lang/Integer.html#Integer-int-