docs: fix documentation
In DeveloperGuide, the instructions to add JUnit 4 to class path
in IntelliJ are out of date. Let's fix it.

In UserGuide, the instructions for "Changing of save location" still
use the `.xml` file extension, when it has been updated to `.txt`
in #388. Let's fix it as well.

In LearningOutcome, there are some typos and incorrect links. Let's
fix them as well.