docs: fix documentation
In DeveloperGuide, the instructions to add JUnit 4 to class path
in Intellij is out of date. Let's fix it.

In UserGuide, the instructions for Changing of save location still
uses the `.xml` file extension, when it has been updated to `.txt`
in #388. Let's fix it as well.

In LearningOutcome, there some typos and broken links to the
se-edu/se-book. Let's fix them as well.