Remove Gradle wrapper
Our repo contains the Gradle wrapper to ensure that the correct Gradle
version is used when we are building our documentation with asciidoctor.

However, IntelliJ has a strange quirk where if:

  * The project is imported by "Create Project from existing Sources",
    and,

  * The `gradle-wrapper.jar` in the repo is selected as a library
    (which is done by default)

then IntelliJ might end up marking the `src` directory as "not a sources
root directory", resulting in IntelliJ not being able to compile and run
the project[1].

While workarounds exists (such as telling users not to select
`gradle-wrapper.jar` in the library selection stage), ultimately this
boils down to the fact that we include the gradle wrapper in the repo
even though ordinary users won't be using it.

So, let's remove the Gradle wrapper in the repo.

To ensure that Travis/Netlify install and use the correct Gradle
version, let's set up a system where Travis/Netlify will download the
Gradle wrapper at build time. We are already hosting the Gradle wrapper
of the version we want at https://se-edu.github.io/gradle-wrapper

[1] https://github.com/se-edu/addressbook-level1/commit/13daa30527fb541e6e7d073ec308537d1f21d871