Rename `build.gradle` to `build.gradle.txt`
We have a `build.gradle` file in the repo which tells Travis/Netlify how
to build our documentation with asciidoctor.

However, when importing this repo as a project, IntelliJ will detect the
`build.gradle` file and prompt the user with the following:

    Unlinked Gradle project?

    Import Gradle project, this will also enable Gradle Tool Window...

The correct action that should be taken by the user in this case is
_not_ to "Import Gradle project", as this project is meant to be built
without Gradle. However, this prompt may mislead the user into choosing
the wrong thing.

Fix this by renaming `build.gradle` into `build.gradle.txt`. Keeping the
term `build.gradle` in the filename will allow readers to easily
identify the file as a `build.gradle` file, while the `.txt` extension
will make IntelliJ, and likely all other IDEs, stop identifying it as a
Gradle file (since text files, by definition, can contain anything).