ParserTest: Fix incorrectly formatted command string
Currently, the test methods
parse_addCommandInvalidPersonDataInArgs_errorMessge() and
parse_addCommandInvalidArgs_errorMessage() use an incorrectly formatted
command string. They attempt to format commands with `$s`.

This `$s` is not recognized by Java’s String#format(…), thus the
command is not formatted as expected, but instead simply left untouched.
The tests currently pass, however, because `$s` achieves the purpose of
being an invalid command string. However, this is clearly not the
intention of the test case, which is designed to simulate with sample
names, phones and emails.

To fix the bug in the test, let’s change the occurrences of `$s` to `%s`.