Added ThenListCommand class
This class takes a command and changes its execution behaviour, such that
a ListCommand is executed immediately after the given command is executed.

The ListCommand is executed using the addressbook data that has been modified by the
first Command.
In this way, the ThenListCommand can be used when we want to perform a Command, and then
immediately list the results of the address book.