Utils#isAnyNull: Check for null on entire object
Currently, isAnyNull assumes that the `items` varargs is not null.

However, this is not necessarily the case. It can be the case that
`isAnyNull(null)` is called. If the varargs is null, the function should
return false, since there is nothing to be examined. Presently, the code
will simply crash with a NullPointerException if null is passed in as an
argument.

Let's add checks to ensure that the varargs object is not null.