Migrate to JUnit 5 from JUnit 4
Change imports to JUnit 5 imports.

Replace @Before with @BeforeEach, ExpectedException with assertThrows,
TemporaryFolder with TempDir where applicable as the former are
removed in JUnit 5.

Add dependencies for JUnit 5 to build.gradle.txt.

Update DG to import JUnit 5.4x.

Add images to DG to illustrate the process of downloading of JUnit5.4
from Maven as the process varies on different versions of IntelliJ IDEA.

Update the DG to reflect the fully qualified package name of JUnit5.