Migrate to checkstyle 8.2
Upgrade checkstyle version and checkstyle rules to match that of later
AddressBooks for consistency.

Add checkstyle checks into the buildfile to ensure that future commits
will remain compliant to checkstyle rules.

Fix checkstyle violations as the AddressBooks are meant to be models for
students to learn from.

Modify checkstyle rules to allow JUnit4 annotations until the migration
to JUnit5.