Update checkstyle ruleset to match later levels of AddressBook.
Upgrade checkstyle version to 8.1 and checkstyle rules to match that of
later AddressBooks for consistency.

Add checkstyle checks into the buildfile to ensure that future commits
will remain compliant to checkstyle rules.

Fix checkstyle violations as the AddressBooks are meant to be models for
students to learn from.

Update dependencies in build.gradle.txt to reflect the required version
of checkstyle.

Add checkstyle checks to travis build.